# docs/openapi.yaml
openapi: 3.0.3
info:
  title: UAV Platform API
  version: 2.0.0
  description: |
    基于WRF气象驱动的无人机VRP智能路径规划系统 API 平台。

    ## 功能模块
    - **平台管理**: 租户管理、API密钥管理、用量统计
    - **气象数据**: WRF/风云/天姿/风雷多源气象数据查询与融合
    - **数据同化**: 3DVAR/4DVAR/5DVAR/EnKF/Hybrid/Enhanced Bayesian 同化算法
    - **风险评估**: 综合风险感知与区域风险栅格地图
    - **适航评估**: 全维度适航评估与适航标准查询
    - **观测决策**: 自适应/计划/应急观测任务管理与决策
    - **路径规划**: VRPTW/DE-RRT*/DWA/MPC/A*/Dijkstra/RRT* 路径规划
    - **UTM**: 空域管理、飞行计划、实时追踪与冲突检测

    ## 认证方式
    - **Bearer Token**: JWT 认证，用于平台管理接口
    - **API Key**: HMAC 签名认证，用于业务接口调用

  contact:
    name: UAV Platform Team
    email: support@uav-platform.example.com
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: http://localhost:8080
    description: 开发环境（通过网关访问）
  - url: https://api.uav-platform.example.com
    description: 生产环境
...
paths:
...
  /api/v1/platform/tenants:
    post:
      tags: [Platform]
      summary: 创建租户
      description: 创建新的租户账户，包含名称、数据库Schema和配额配置。
      operationId: createTenant
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTenantRequest'
            example:
              name: "示例航空公司"
              schemaName: "tenant_airline_01"
              quotaConfig: '{"maxUavs": 50, "maxApiCallsPerDay": 10000}'
      responses:
        '200':
          description: 租户创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultTenant'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'

    get:
      tags: [Platform]
      summary: 分页查询租户列表
      description: 分页获取所有租户信息。
      operationId: listTenants
      security:
        - BearerAuth: []
      parameters:
        - name: current
          in: query
          description: 当前页码
          required: false
          schema:
            type: integer
            default: 1
        - name: size
          in: query
          description: 每页大小
          required: false
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultPageTenant'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/platform/tenants/{id}:
    get:
      tags: [Platform]
      summary: 获取租户详情
      description: 根据ID获取租户详细信息。
      operationId: getTenant
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TenantId'
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultTenant'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags: [Platform]
      summary: 更新租户信息
      description: 更新指定租户的名称和配额配置。
      operationId: updateTenant
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TenantId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTenantRequest'
      responses:
        '200':
          $ref: '#/components/responses/Success'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Platform]
      summary: 删除租户
      description: 删除指定租户及其所有关联数据。
      operationId: deleteTenant
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TenantId'
      responses:
        '200':
          $ref: '#/components/responses/Success'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/platform/tenants/{id}/disable:
    post:
      tags: [Platform]
      summary: 禁用租户
      description: 禁用指定租户，禁用后该租户无法访问任何API。
      operationId: disableTenant
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TenantId'
      responses:
        '200':
          $ref: '#/components/responses/Success'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/platform/tenants/{id}/enable:
    post:
      tags: [Platform]
      summary: 启用租户
      description: 重新启用已禁用的租户。
      operationId: enableTenant
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TenantId'
      responses:
        '200':
          $ref: '#/components/responses/Success'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/v1/platform/api-keys/tenant/{tenantId}:
    get:
      tags: [Platform]
      summary: 查询租户的API密钥列表
      description: 获取指定租户下所有API密钥。
      operationId: listApiKeysByTenant
      security:
        - BearerAuth: []
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultListApiKey'
  /api/v1/platform/usage/daily:
    get:
      tags: [Platform]
      summary: 每日用量聚合
      description: 查询指定租户在时间范围内的每日API调用用量统计。
      operationId: getDailyUsage
      security:
        - ApiKeyAuth: []
      parameters:
        - name: tenantId
          in: query
          required: true
          schema:
            type: integer
            format: int64
        - name: startDate
          in: query
          required: true
          description: 开始日期 (yyyy-MM-dd)
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          required: true
          description: 结束日期 (yyyy-MM-dd)
          schema:
            type: string
            format: date
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultListMap'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/v1/platform/usage/api-path:
    get:
      tags: [Platform]
      summary: API路径用量聚合
      description: 按API路径维度聚合用量统计。
      operationId: getApiPathUsage
      security:
        - ApiKeyAuth: []
      parameters:
        - name: tenantId
          in: query
          required: true
          schema:
            type: integer
            format: int64
        - name: startDate
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          required: true
          schema:
            type: string
            format: date
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultListMap'

  # ============================================================
  # 气象数据
  # ============================================================
...
  /api/v1/risk/assess:
    post:
      tags: [Risk]
      summary: 综合风险评估
      description: 对指定位置进行综合风险评估，考虑气象、地形、空域、设备等多维度风险因子。
      operationId: assessRisk
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RiskQueryRequest'
            example:
              longitude: 116.4074
              latitude: 39.9042
              altitude: 100.0
              uavModel: "DJI-M300"
              missionType: "INSPECTION"
      responses:
        '200':
          description: 评估成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultRiskAssessment'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/v1/risk/map:
    get:
      tags: [Risk]
      summary: 区域风险栅格地图
      description: 生成指定矩形区域的风险栅格地图，用于可视化展示风险分布。
      operationId: generateRiskMap
      security:
        - ApiKeyAuth: []
      parameters:
        - name: minLon
          in: query
          required: true
          schema:
            type: number
            format: double
        - name: minLat
          in: query
          required: true
          schema:
            type: number
            format: double
        - name: maxLon
          in: query
          required: true
          schema:
            type: number
            format: double
        - name: maxLat
          in: query
          required: true
          schema:
            type: number
            format: double
        - name: resolution
          in: query
          required: false
          description: 栅格分辨率(度)
          schema:
            type: number
            format: double
            default: 0.01
      responses:
        '200':
          description: 生成成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultListRiskAssessment'
  /api/v1/risk/history:
    get:
      tags: [Risk]
      summary: 历史风险评估记录
      description: 查询历史风险评估记录，支持按租户和风险类型过滤。
      operationId: getRiskHistory
      security:
        - ApiKeyAuth: []
      parameters:
        - name: tenantId
          in: query
          required: false
          schema:
            type: integer
            format: int64
        - name: type
          in: query
          required: false
          description: 风险类型 (WEATHER/TERRAIN/AIRSPACE/EQUIPMENT/COMPOSITE)
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultListRiskAssessment'

  # ============================================================
  # 适航评估
  # ============================================================
...
  /api/v1/airspaces:
    get:
      tags: [UTM]
      summary: 查询空域列表
      description: 获取所有空域信息，包括静态空域、动态空域和限制空域。
      operationId: getAirspaces
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultListAirspace'

    post:
      tags: [UTM]
      summary: 创建动态空域
      description: 创建新的动态空域限制区域。
      operationId: createAirspace
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Airspace'
            example:
              type: "DYNAMIC"
              boundsJson: '{"type":"Polygon","coordinates":[...]}'
              altitudeMin: 0
              altitudeMax: 300
              status: "ACTIVE"
              tenantId: "1"
      responses:
        '200':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultAirspace'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/v1/airspaces/check:
    get:
      tags: [UTM]
      summary: 空域限制检查
      description: 检查指定坐标和高度是否受到空域限制。
      operationId: checkAirspaceRestriction
      security:
        - ApiKeyAuth: []
      parameters:
        - name: lon
          in: query
          required: true
          schema:
            type: number
            format: double
        - name: lat
          in: query
          required: true
          schema:
            type: number
            format: double
        - name: altitude
          in: query
          required: true
          schema:
            type: number
            format: double
      responses:
        '200':
          description: 检查成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultBoolean'
        '400':
          $ref: '#/components/responses/BadRequest'

  # ============================================================
  # UTM - 飞行计划
  # ============================================================
...
components:
...
  parameters:
...
    TenantId:
      name: id
      in: path
      required: true
      description: 租户ID
      schema:
        type: integer
        format: int64
...
  schemas:
...
    Result:
      type: object
      description: 统一API响应结构
      properties:
        code:
          type: integer
          description: 业务状态码 (200=成功)
          example: 200
        message:
          type: string
          description: 提示消息
          example: success
        data:
          description: 响应数据（泛型）
        requestId:
          type: string
          description: 请求追踪ID
        timestamp:
          type: integer
          format: int64
          description: 时间戳
    ResultVoid:
      allOf:
        - $ref: '#/components/schemas/Result'
      description: 无数据的成功响应
...
    ResultBoolean:
      allOf:
        - $ref: '#/components/schemas/Result'
      description: 包含布尔值的响应
      properties:
        data:
          type: boolean
          description: 检查结果
...
    ResultListMap:
      allOf:
        - $ref: '#/components/schemas/Result'
      description: 包含Map列表的响应
      properties:
        data:
          type: array
          items:
            type: object
            additionalProperties:
              type: object

    # --- 平台管理 ---
    CreateTenantRequest:
      type: object
      required: [name, schemaName]
      properties:
        name:
          type: string
          description: 租户名称
          example: "示例航空公司"
        schemaName:
          type: string
          description: 数据库Schema名称
          example: "tenant_airline_01"
        quotaConfig:
          type: string
          description: 配额配置JSON
          example: '{"maxUavs": 50, "maxApiCallsPerDay": 10000}'
    UpdateTenantRequest:
      type: object
      properties:
        name:
          type: string
          description: 租户名称
        quotaConfig:
          type: string
          description: 配额配置JSON
    Tenant:
      type: object
      description: 租户实体
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        schemaName:
          type: string
        status:
          type: integer
          description: "状态: 1=启用, 0=禁用"
        quotaConfig:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ResultTenant:
      allOf:
        - $ref: '#/components/schemas/Result'
      description: 包含租户数据的响应
      properties:
        data:
          $ref: '#/components/schemas/Tenant'
    ResultPageTenant:
      allOf:
        - $ref: '#/components/schemas/Result'
      description: 包含租户分页数据的响应
      properties:
        data:
          type: object
          properties:
            records:
              type: array
              items:
                $ref: '#/components/schemas/Tenant'
            total:
              type: integer
            size:
              type: integer
            current:
              type: integer
...
    ApiKey:
      type: object
      description: API密钥实体
      properties:
        id:
          type: integer
          format: int64
        tenantId:
          type: integer
          format: int64
        keyValue:
          type: string
          description: API Key值
        secret:
          type: string
          description: 密钥Secret
        name:
          type: string
        status:
          type: integer
          description: "状态: 1=启用, 0=禁用"
        rateLimit:
          type: integer
        createdAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
...
    ResultListApiKey:
      allOf:
        - $ref: '#/components/schemas/Result'
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/ApiKey'

    # --- 气象数据 ---
...
    RiskQueryRequest:
      type: object
      required: [longitude, latitude]
      properties:
        longitude:
          type: number
          format: double
        latitude:
          type: number
          format: double
        altitude:
          type: number
          format: double
          description: 海拔高度(m)
        uavModel:
          type: string
          description: 无人机型号
        missionType:
          type: string
          description: 任务类型
    RiskAssessment:
      type: object
      description: 风险评估结果
      properties:
        id:
          type: integer
          format: int64
        type:
          type: string
          description: "风险类型: WEATHER, TERRAIN, AIRSPACE, EQUIPMENT, COMPOSITE"
        level:
          type: integer
          description: 风险等级(1-5)
        score:
          type: integer
          description: 风险评分(0-100)
        factorsJson:
          type: string
          description: 风险因子JSON
        locationJson:
          type: string
          description: 位置信息JSON
        tenantId:
          type: integer
          format: int64
        createdAt:
          type: string
          format: date-time
    ResultRiskAssessment:
      allOf:
        - $ref: '#/components/schemas/Result'
      properties:
        data:
          $ref: '#/components/schemas/RiskAssessment'
    ResultListRiskAssessment:
      allOf:
        - $ref: '#/components/schemas/Result'
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/RiskAssessment'

    # --- 适航评估 ---
...
    Airspace:
      type: object
      description: 空域信息
      properties:
        id:
          type: integer
          format: int64
        type:
          type: string
          enum: [STATIC, DYNAMIC, RESTRICTED]
          description: 空域类型
        boundsJson:
          type: string
          description: 边界JSON
        altitudeMin:
          type: number
          format: double
          description: 最低高度(m)
        altitudeMax:
          type: number
          format: double
          description: 最高高度(m)
        effectiveTimeStart:
          type: string
          format: date-time
          description: 生效开始时间
        effectiveTimeEnd:
          type: string
          format: date-time
          description: 生效结束时间
        status:
          type: string
          enum: [ACTIVE, INACTIVE]
          description: 空域状态
        tenantId:
          type: string
    ResultAirspace:
      allOf:
        - $ref: '#/components/schemas/Result'
      properties:
        data:
          $ref: '#/components/schemas/Airspace'
    ResultListAirspace:
      allOf:
        - $ref: '#/components/schemas/Result'
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Airspace'
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Token 认证，用于平台管理接口

    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API Key 认证，配合 HMAC 签名使用

  # ----------------------------------------------------------
  # 通用响应
  # ----------------------------------------------------------
  responses:
    Success:
      description: 操作成功
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ResultVoid'
          example:
            code: 200
            message: success
            data: null
            timestamp: 1717200000000
    BadRequest:
      description: 请求参数错误
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Result'
          example:
            code: 1000
            message: "请求参数错误"
            data: null
            timestamp: 1717200000000
    Unauthorized:
      description: 未授权，请先登录
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Result'
          example:
            code: 1001
            message: "未授权，请先登录"
            data: null
            timestamp: 1717200000000
    Forbidden:
      description: 权限不足
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Result'
          example:
            code: 1002
            message: "权限不足"
            data: null
            timestamp: 1717200000000
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Result'
          example:
            code: 1003
            message: "资源不存在"
            data: null
            timestamp: 1717200000000
    InternalServerError:
      description: 服务器内部错误
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Result'
          example:
            code: 2000
            message: "服务器内部错误"
            data: null
            timestamp: 1717200000000